+++ entity-view/processor/src/main/java/com/blazebit/persistence/view/processor/AttributeFilter.java
@@ -16,7 +16,6 @@
 
 package com.blazebit.persistence.view.processor;
 
-import javax.lang.model.element.TypeElement;
 import javax.lang.model.type.DeclaredType;
 import javax.lang.model.type.TypeKind;
 import javax.lang.model.type.TypeMirror;
@@ -30,11 +29,11 @@ public class AttributeFilter {
     private final String name;
     private final JavaType filterValueType;
 
-    public AttributeFilter(String name, TypeElement filterProvider, String attributeType, Context context) {
+    public AttributeFilter(String name, TypeMirror filterProvider, String attributeType, Context context) {
         this.name = name;
-        TypeMirror typeMirror = TypeUtils.asMemberOf(context, (DeclaredType) filterProvider.asType(), context.getTypeElement(Constants.ATTRIBUTE_FILTER_PROVIDER).getTypeParameters().get(0));
+        TypeMirror typeMirror = TypeUtils.asMemberOf(context, (DeclaredType) filterProvider, context.getTypeElement(Constants.ATTRIBUTE_FILTER_PROVIDER).getTypeParameters().get(0));
 
-        if (typeMirror.getKind() == TypeKind.TYPEVAR) {
+        if (typeMirror.getKind() == TypeKind.TYPEVAR || "java.lang.Object".equals(typeMirror.toString())) {
             this.filterValueType = new JavaType(attributeType);
         } else {
             this.filterValueType = new JavaType(typeMirror);